feat: sse 기반 알림 메시지 조회, 읽음 처리 로직 작성

+++ src/main/java/com/mincho/herb/common/dto/PageInfoDTO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Builder
 public class PageInfoDTO {
-    private Long page;
-    private Long size;
+    private int page;
+    private int size;
 }
